Use SQLITE3_LIBNAME to set the linked library
This change updates the build script for libsqlcipher-sys.  Before this
change, the build script always linked "sqlite3".  With this change,
setting SQLITE3_LIBNAME will cause libsqlcipher-sys to link a different
library.  This is useful on systems where both libsqlite3 and
libsqlcipher are installed, where it may be desirable to use
SQLITE3_LIBNAME=sqlcipher when building.